+++ src/Components/NoticeCheck/NoticeCheck.tsx
@@ -3,11 +3,10 @@ import * as S from './Style';
 import * as I from '../../Assets/Svg/index';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useRole } from 'Hooks/useRole';
-import { getNoticeItem } from 'Api/notice';
+import { getNoticeItem, noticeUpdate } from 'Api/notice';
 
 const NoticePage: React.FC = () => {
-    const {board_key} = useParams();
-    
+    const {board_key}:any = useParams();
 	const [title, setTitle] = useState<string>('');
 	const [content, setContent] = useState<string>('');
     const [imgUrl, setImgUrl] = useState<string>('');
@@ -15,6 +14,7 @@ const NoticePage: React.FC = () => {
     const [roles, setRoles] = useState<string>('');
     const role = useRole();
     const navigate = useNavigate();
+    const [updateState, setUpdateState] = useState<boolean>(false);
 
 	const GetNoticeItem = async (boardId) => {
 		return getNoticeItem(role, boardId);
@@ -39,29 +39,65 @@ const NoticePage: React.FC = () => {
                 setContent(res.content);
                 setDate(res.createdDate);
                 setRoles(res.roles);
-				// setUpdateTitle(res.title);
-				// setUpdateContent(res.content);
+				setUpdateTitle(res.title);
+				setUpdateContent(res.content);
             	setImgUrl(res.url);
 			});
-    }, []);    
+    }, []);
+    const [updateTitle, setUpdateTitle] = useState<string>('');
+	const [updateContent, setUpdateContent] = useState<string>('');
+
+	const onUpdate = async (e) => {
+		await noticeUpdate(role, board_key, updateTitle, updateContent);
+		setUpdateState(false);
+		window.location.reload();
+	};
     
 	return (
 		<S.Positioner>
             <S.Container>
                 <S.HeaderWrapper>
                     <S.Header>
-                        <S.TitleWrapper>{title}</S.TitleWrapper>
+                        {updateState ? 
+							<input
+                                value={updateTitle}
+                                onChange={(e) => setUpdateTitle(e.target.value)}
+                            />                        
+                        :
+                            <S.TitleWrapper>{title}</S.TitleWrapper>
+                        }
                         <S.DateWrapper>{date.substr(0, 10)}</S.DateWrapper>
                     </S.Header>
                     <S.CancelWrapper><I.Cancel onClick={() => navigate('/notice')}/></S.CancelWrapper>
                 </S.HeaderWrapper>
                 <S.ContentWrapper>
-                    <S.TextWrapper>{content}</S.TextWrapper>
+                    {updateState ? 
+                        <S.TextWrapper>
+                            <textarea
+                                value={updateContent}
+                                onChange={(e) => setUpdateContent(e.target.value)}
+                            />
+                        </S.TextWrapper>
+                    :
+                        <S.TextWrapper>
+							{content?.split('\n').map((line) => (
+							    <p>
+								    {line}
+								    <br />
+						        </p>
+					        ))}                            
+                        </S.TextWrapper>
+                    }
                     {imgUrl && (
                         <S.ImgWrapper alt="notice" src={imgUrl}/>
                     )}
                 </S.ContentWrapper>
                 <S.Footer>작성자 : {returnAuthorValue(roles[0])}</S.Footer>
+                {updateState ? 
+                <S.EditBtn onClick={onUpdate}>확인</S.EditBtn>
+                :
+                <S.EditBtn onClick={() => {setUpdateState(true)}}>수정</S.EditBtn>    
+            }
             </S.Container>
 		</S.Positioner>
 	);
